+++ libs/docs/platform/table/examples/platform-table-multiple-row-selection-example.component.ts
@@ -36,6 +36,8 @@ export interface ExampleItem {
     statusColor?: string;
     date: FdDate;
     verified: boolean;
+    selectable?: boolean;
+    selected?: boolean;
 }
 
 /**
@@ -93,7 +95,8 @@ const ITEMS: ExampleItem[] = [
         status: 'Stocked on demand',
         statusColor: 'informative',
         date: new FdDate(2020, 1, 7),
-        verified: true
+        verified: true,
+        selectable: false
     },
     {
         name: 'Astro Laptop 1516',
@@ -105,7 +108,8 @@ const ITEMS: ExampleItem[] = [
         status: 'Out of stock',
         statusColor: 'negative',
         date: new FdDate(2020, 2, 5),
-        verified: true
+        verified: true,
+        selected: true
     },
     {
         name: 'Astro Phone 6',