+++ libs/platform/src/lib/table/models/index.ts
@@ -12,3 +12,4 @@ export * from './table-rows-rearrange-event.model';
 export * from './table-row-activate-event.model';
 export * from './row-comparator.model';
 export * from './table-managed-preset';
+export * from './save-rows-event.interface';